refactor(libsinsp): move cri settings in another file
